Add replaceCart reducer in the cart-slice to only replace the state.
When this is done the next step is to send the fetch request to the Server and send the Data, but now not directly in the Reducer but outside in the ProductItem Component. This is forbidden to be done directly in the Reducer function because it must be a Pure Function!